bug resolved, task complete

+++ Lab_2_180042119/task/controllers/userController.controllers.js
@@ -3,7 +3,6 @@ const bcrypt = require("bcryptjs");
 var LocalStorage = require("node-localstorage").LocalStorage;
 localStorage = new LocalStorage("./scratch");
 const alert = require("alert");
-const saltRounds = 10;
 
 const getRegister = (req, res) => {
   res.sendFile("register.html", { root: "./views/pages/examples" });
@@ -16,7 +15,11 @@ const postRegister = async (req, res) => {
   const re_pass = req.body.re_pass;
 
   try {
-    if (String(password).length < 6) {
+    userExist = await User.findOne({ email });
+    if (userExist) {
+      alert("There is already an account under that email.");
+      res.redirect("/register");
+    } else if (String(password).length < 6) {
       alert("Please enter a password of at least 6 characters.");
       res.redirect("/register");
     } else if (!(password == re_pass)) {
@@ -32,7 +35,7 @@ const postRegister = async (req, res) => {
         passwordHash,
       });
       const savedUser = await user.save();
-      localStorage.setItem("fullname", name)
+      localStorage.setItem("fullname", name);
       res.cookie("fullname", name);
       res.redirect("/dashboard");
     }
@@ -50,21 +53,20 @@ const postLogin = async (req, res) => {
   const email = req.body.email;
   const pass = req.body.pass;
   const existingUser = await User.findOne({ email });
- try{ if (existingUser) {
-    const passMatch = bcrypt.compare(pass, existingUser.passwordHash);
+  if (existingUser) {
+    const passMatch = await bcrypt.compare(pass, existingUser.passwordHash);
     if (passMatch) {
-      localStorage.setItem("name", existingUser.name)
+      localStorage.setItem("name", existingUser.name);
       // res.cookie("fullname", existingUser.name);
+
       res.redirect("/dashboard");
     } else {
       alert("Wrong Password");
+      res.redirect("/login");
     }
   } else {
     alert("You are not registered\nPlease create an account");
     res.redirect("/login");
-  }}catch (error){
-    console.error(error.message);
-    res.status(500).send(error.message);
   }
 };
 
@@ -72,4 +74,10 @@ const getDashboard = (req, res) => {
   res.sendFile("index3.html", { root: "./views" });
 };
 
-module.exports = { getLogin, getRegister, postRegister, postLogin,getDashboard };
+module.exports = {
+  getLogin,
+  getRegister,
+  postRegister,
+  postLogin,
+  getDashboard,
+};
